third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/Document-importNode-expected.txt

This is a testharness.js-based test.
[FAIL] importNode should clone using the specified registry if target's registry is null
  assert_true: expected true got false
[FAIL] importNode should clone a shadow host with a declarative shadow DOM using the global registry by default
  assert_true: expected true got false
[FAIL] importNode should clone a shadow host with a declarative shadow DOM using a specified scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] importNode should clone using target's registry if non-null, including when it's not the global registry
  assert_false: expected false got true
[FAIL] importNode should clone a template content using the global registry by default
  assert_equals: expected "GlobalSomeElement" but got "HTMLElement"
[FAIL] importNode should clone a template content using a specified scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] importNode should clone a template content with a nested template element using a scoped registry
  assert_equals: expected "GlobalSomeElement" but got "HTMLElement"
Harness: the test ran to completion.
